+++ packages/extensions/core/src/lib/plugins/enum-deduplication/enum-deduplicator.ts
@@ -30,8 +30,7 @@ export class EnumDeduplicator extends TransformerViaPointer<oai3.Model, oai3.Mod
         return false;
       }
       // use the given name if specified, otherwise fallback to the metadata name
-      const name = pascalCase(value["x-ms-enum"] ? value["x-ms-enum"].name : value["x-ms-metadata"].name);
-
+      const name = pascalCase(getEnumName(value));
       const e = this.enums.get(name) || this.enums.set(name, []).get(name) || [];
       e.push({ target, value, key, pointer, originalNodes });
       return true;
@@ -51,7 +50,7 @@ export class EnumDeduplicator extends TransformerViaPointer<oai3.Model, oai3.Mod
       );
 
       const first = enumSet[0];
-      const name = first.value["x-ms-enum"] ? first.value["x-ms-enum"].name : first.value["x-ms-metadata"].name;
+      const name = getEnumName(first.value);
       if (enumSet.length === 1) {
         const originalRef = `#/components/schemas/${first.key}`;
         const newRef = `#/components/schemas/${name}`;
@@ -70,10 +69,15 @@ export class EnumDeduplicator extends TransformerViaPointer<oai3.Model, oai3.Mod
       if (first.value.description) {
         this.clone(mergedEnum, "description", first.pointer, first.value.description);
       }
+      if (first.value.type) {
+        this.clone(mergedEnum, "type", first.pointer, first.value.type);
+      }
       if (first.value["x-ms-enum"]) {
         this.clone(mergedEnum, "x-ms-enum", first.pointer, first.value["x-ms-enum"]);
       }
-      this.clone(mergedEnum, "type", first.pointer, "string");
+      if (first.value["format"]) {
+        this.clone(mergedEnum, "format", first.pointer, first.value["format"]);
+      }
       const newRef = `#/components/schemas/${name}`;
       this.newArray(mergedEnum, "enum", "");
 
@@ -116,4 +120,13 @@ export class EnumDeduplicator extends TransformerViaPointer<oai3.Model, oai3.Mod
       }
     }
   }
+}
+
+/**
+ * Returns the name of the enum. Either provided via `x-ms-enum.name` or resolved automatically.
+ * @param value Enum Schema
+ * @returns name of the enum.
+ */
+function getEnumName(value: oai3.Schema): string {
+  return value["x-ms-enum"]?.name ? value["x-ms-enum"].name : value["x-ms-metadata"].name;
 }