+++ packages/tools/fixer/src/app/autorest-fixer-config.ts
@@ -1,0 +1,8 @@
+export interface AutorestFixerConfig {
+  include: string[];
+
+  /**
+   * If the fixer should just do a dry run.
+   */
+  dryRun?: boolean;
+}