+++ packages/libs/configuration/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @autorest/configuration
 
-This log was last generated on Wed, 08 Sep 2021 15:39:22 GMT and should not be manually modified.
+This log was last generated on Thu, 23 Sep 2021 19:51:32 GMT and should not be manually modified.
+
+## 1.7.2
+Thu, 23 Sep 2021 19:51:32 GMT
+
+### Patches
+
+- **Added** `include-x-ms-examples-original-file` flag to activate `x-ms-original-file` injection in `x-ms-examples`
+
+## 1.7.1
+Wed, 22 Sep 2021 15:23:39 GMT
+
+### Patches
+
+- Added missing help text for command line arguments
 
 ## 1.7.0
 Wed, 08 Sep 2021 15:39:22 GMT